Fix importers to create editions that have lead organisations by supplying lead_edition_organisations attributes with EditionOrganisation instances in them, rather than poking directly onto the organisations association.
In passing we add a before_validation on Editions that makes sure that all of its EditionOrganistion instances have an edition instance that points to itself.  This gets around the case where you have a Edition instance and you try to add an EditionOrganisation instance that you built before you had access to the Edition and thus the EditionOrganisation fails validation because it has no Edition (e.g. exactly what happens in the importers).