Extract edition_organisations_attributes fields into a helper on the edition_form_builder.
Also because we submit our changes to edition_organisations_attributs we should look to see if the object in the form has errors.  If it does we should render the fields based on edition_organisations rather than lead_edition_organisations or supporting_edition_organisations, if it doesn't we can use the simpler API of lead_edition_organisations or supporting_edition_organisations.  We then have to do the same thing in build_edition_organisations too to make sure that we have enough "empty" objects lying around to present to the user.